[#3545] sticky elements
