+++ akvo/rsr/spa/app/components/editor/styles.scss
@@ -1,10 +1,51 @@
 @import url(https://fonts.googleapis.com/css?family=Lato:300&subset=latin);
 
 .editor{
-  display: flex;
-  flex-direction: row;
+  
+  // padding-top: 50px;
+  position: relative;
+  .flex-container{
+    display: flex;
+    flex-direction: row;
+  }
+  .status-bar{
+    position: sticky;
+    height: 50px;
+    background-color: #f2f2f2;
+    width: 100%;
+    top: 0;
+    display: flex;
+    line-height: 50px;
+    z-index: 2;
+    aside{
+      background: #e6e6e6;
+      height: 50px;
+      padding: 0 20px;
+      .anticon{
+        margin-right: 10px;
+        color: green;
+      }
+      span{
+        font-size: 14px;
+        opacity: 0.5;
+      }
+    }
+    .content{
+      display: flex;
+      flex-direction: row-reverse;
+      i{
+        opacity: 0.5;
+      }
+      .ant-btn{
+        margin-top: 9px;
+        margin-left: 10px;
+        margin-right: 10px;
+      }
+    }
+  }
   aside{
-    width: 270px;
+    width: 270px!important;
+    position: absolute;
     ul{
       background-color: #f2f2f2;
       padding: 20px;
@@ -29,11 +70,18 @@
   }
   .content{
     flex: 1;
+  }
+  &> .flex-container > .content{
     padding-left: 30px;
+    margin-left: 270px;
+    margin-right: 330px;
   }
   .alerts{
     width: 300px;
-    padding-left: 30px;
+    position: absolute;
+    right: 0;
+    // padding-left: 30px;
+    padding-top: 20px;
   }
 }
 
@@ -44,9 +92,12 @@ hr{
   margin: 20px 0;
 }
 
+.view {
+  padding-top: 20px;
+}
+
 .settings.view {
   p:first-of-type{
-    margin-top: 20px;
     margin-bottom: 20px;
   }
   .switch-label{
