+++ akvo/rsr/spa/app/root.js
@@ -5,7 +5,7 @@ import 'reset-css'
 import 'antd/dist/antd.css'
 
 import TopBar from './components/top-bar'
-import Editor from './components/editor/comp/root'
+import Editor from './components/editor/comp'
 
 const App = () => (
   <div>Hello world</div>
