+++ akvo/rsr/spa/app/styles/main.scss
@@ -6,19 +6,20 @@ body{
 
 .top-bar{
   padding: 20px 0;
-  height: 80px;
-  background-color: #f2f2f2;
+  height: 70px;
+  // background-color: #f2f2f2;
 }
 
 header{
   border-bottom: 1px solid #ccc;
-  padding: 15px 0;
+  padding: 10px 0;
   h1{
     font-weight: 300;
-    font-size: 24px;
+    font-size: 21px;
     color: #2F1A45;
+    margin-bottom: 0;
   }
-  margin-bottom: 30px;
+  margin-bottom: 10px;
 }
 
 @import "./container.scss"
